# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Nafex Hub API - Fashion marketplace platform for Ghana
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/paystack/initialize:
    post:
      operationId: initializePaystackPayment
      tags: [payments]
      summary: Initialize Paystack payment for an order
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentBody"
      responses:
        "200":
          description: Paystack authorization URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaystackInitResponse"
        "503":
          description: Payment gateway not configured
  /payments/paystack/verify:
    post:
      operationId: verifyPaystackPayment
      tags: [payments]
      summary: Verify Paystack payment and lock escrow
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: Verification result and updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaystackVerifyResponse"
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [payments]
      summary: Get user transaction history
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /admin/transactions:
    get:
      operationId: getAdminTransactions
      tags: [admin, payments]
      summary: All transactions (admin)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /admin/payouts/{orderId}:
    post:
      operationId: releasePayout
      tags: [admin, payments]
      summary: Release escrow payout to seller (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated order
        "409":
          description: Order not in escrow
  /admin/refunds/{orderId}:
    post:
      operationId: processRefund
      tags: [admin, payments]
      summary: Process refund for an order (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Updated order
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
          nullable: true
        userId:
          type: integer
        type:
          type: string
          enum: [payment, refund, payout, fee]
        amount:
          type: string
        currency:
          type: string
        provider:
          type: string
          enum: [paystack, momo, manual, system]
        providerRef:
          type: string
          nullable: true
        channel:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, success, failed, reversed]
        metadata:
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, amount, currency, provider, status, createdAt]
    InitiatePaymentBody:
      type: object
      properties:
        orderId:
          type: integer
        channel:
          type: string
          enum: [card, mobile_money, bank]
        momoPhone:
          type: string
        momoNetwork:
          type: string
          enum: [MTN, Vodafone, AirtelTigo]
      required: [orderId]
    PaystackInitResponse:
      type: object
      properties:
        authorizationUrl:
          type: string
        accessCode:
          type: string
        reference:
          type: string
      required: [authorizationUrl, accessCode, reference]
    VerifyPaymentBody:
      type: object
      properties:
        reference:
          type: string
        orderId:
          type: integer
      required: [reference, orderId]
    PaystackVerifyResponse:
      type: object
      properties:
        order:
          type: object
        transaction:
          type: object
      required: [order, transaction]
...